Include file path in error message
This is especially helpful when the file listed in the "extends"
attribute cannot be read. Before this change, the path listed was the
outermost tsconfig.json, which can be very confusing.

In order for this to work, an additional error type had to be created
to distinguish between the two io:Error occurrences. This should also
help to narrow down errors.